Factoring private type buffer out of cell.
