+++ cell/color_test.go
@@ -14,7 +14,10 @@
 
 package cell
 
-import "testing"
+import (
+	"fmt"
+	"testing"
+)
 
 func TestColorNumber(t *testing.T) {
 	tests := []struct {
@@ -46,6 +49,7 @@ func TestColorNumber(t *testing.T) {
 
 	for _, tc := range tests {
 		t.Run(tc.desc, func(t *testing.T) {
+			t.Logf(fmt.Sprintf("color: %v", tc.want))
 			got := ColorNumber(tc.number)
 			if got != tc.want {
 				t.Errorf("ColorNumber(%v) => %v, want %v", tc.number, got, tc.want)
